refactor(mesh/media): turn api request into actions
