+++ packages/mesh/app/actions/get-latest-stories-in-categroy.ts
@@ -1,4 +1,4 @@
-import 'server-only'
+'use server'
 
 import { RESTFUL_ENDPOINTS } from '@/constants/config'
 import { type LatestStoriesQuery } from '@/graphql/__generated__/graphql'
